removed useless function and moved policy definition
